I have no idea what's happening.
Attempted to make some functions more general (I wonder if type linting makes it run slower...if it does I'll have to remove all of it in a production version.) (Also wonder if a generator to a list is slower than a list comprehension...quite possibly...)

Made additional helper function to test if a group of functions are all above some func id.

Modified high_ord_comp_check to return a set without converting to a list.

The algorithm clearly needs more work.